Allow overriding UID:GID to avoid permission errors
The wordpress and wpcli containers runs as user www-data:www-data (33:33)
by default, but this often does not match the UID:GID of the user on the
host system, and Simply Static would fail to write to the destination
directories unless "sudo chown -R 33:33" was run beforehand, but that is
confusing and cumbersome.  Instead, now we can simply use

    export DOCKER_WP_UID_GID="$(id -u):$(id -g)"

without worrying about directory permissions.